refactor authcontext
