+++ src/front-end/src/components/Login.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 import { connect, ConnectedProps } from "react-redux";
 import { RouteComponentProps } from "react-router-dom";
-import { AuthContext } from "../globalStates/AuthContext";
+import { AuthContext } from "./AuthContext";
 import logo from "../logo.svg";
 import { freshLogin } from "../useCases/logIn/freshLogin";
 
-export interface LoginProps extends RouteComponentProps { }
+export interface LoginProps extends RouteComponentProps {}
 type props = PropsFromRedux & LoginProps;
 export interface LoginState {
   email: string;
@@ -118,7 +118,6 @@ class Login extends React.Component<props, LoginState> {
     );
   }
   async loginDetailsAreCorrect(): Promise<boolean> {
-
     try {
       await freshLogin({
         email: this.state.email,
@@ -131,7 +130,6 @@ class Login extends React.Component<props, LoginState> {
         alert(error.message || "Connection error. Check your internet");
       return false;
     }
-
   }
 }
